Nessus: Added support for api_key (#1183)
* Added support for api_key

* Fixed CR comment

+++ adapters/nessus_adapter/connection.py
@@ -20,6 +20,12 @@ HOST_DETAILS_ENDPOINT = "scans/{0}/hosts/{1}"
 
 class NessusConnection(object):
     def __init__(self, host, port, verify_ssl):
+        self.username = None
+        self.password = None
+        self.access_key = None
+        self.secret_key = None
+        self.token = None
+        self.should_use_token = True  # Setting a default of true
         self.host = host
         self.port = port if port is not None else DEFAULT_NESSUS_PORT
         self.url = host + ':' + str(self.port)
@@ -39,7 +45,7 @@ class NessusConnection(object):
         self.disconnect()
 
     def __del__(self):
-        if hasattr(self, SESSION) and self._is_connected:
+        if hasattr(self, SESSION) and self._is_connected and self.should_use_token:
             try:
                 self.disconnect()
             except Exception:
@@ -50,15 +56,27 @@ class NessusConnection(object):
     def _is_connected(self):
         return self.session is not None
 
-    def set_credentials(self, username, password):
+    def set_credentials(self, username=None, password=None, api_access_key=None, api_secret_key=None):
         """
-
         :param username:
         :param password:
+        :param api_access_key:
+        :param api_secret_key:
         :return:
         """
-        self.username = username
-        self.password = password
+        if username is not None and username != "" and password is not None and password != "":
+            # We should use the user and password
+            self.should_use_token = True
+            self.username = username
+            self.password = password
+        elif api_access_key is not None and api_access_key != "" and \
+                api_secret_key is not None and api_secret_key != "":
+            # We should just use the given api keys
+            self.should_use_token = False
+            self.access_key = api_access_key
+            self.secret_key = api_secret_key
+        else:
+            raise NessusCredentialMissing("Missing user/password or api keys")
 
     def connect(self):
         """
@@ -66,31 +84,38 @@ class NessusConnection(object):
         Save returned token for performing requests, since they require authorization
 
         """
-        if self._is_connected:
+        if self._is_connected and self.should_use_token:
             raise NessusAlreadyConnected()
-        session = requests.Session()
-        if (self.username is None or self.password is None) and self.token is None:
-            raise NessusCredentialMissing()
-        elif self.username is None or self.password is None:
-            return
+        if self.should_use_token:
+            # This code will generate token and create session if user entered user/password
+            session = requests.Session()
+            if (self.username is None or self.password is None) and self.token is None:
+                raise NessusCredentialMissing()
+            elif self.username is None or self.password is None:
+                return
 
-        try:
-            response = session.post(self._get_url_request(SESSION), headers=self.headers,
-                                    json={USERNAME: self.username, PASSWORD: self.password},
-                                    verify=self.verify_ssl, timeout=(5, 30))
-            response.raise_for_status()
-        except requests.HTTPError as e:
-            raise NessusConnectionError(
-                "POST /session with username {0} failed. Details: {1}".format(self.username, str(e)))
-        response = response.json()
-        if TOKEN not in response:
-            error = response.get('errorCode', 'Unknown connection error')
-            message = response.get('errorMessage', '')
-            if message:
-                error = '{0}: {1}'.format(error, message)
-            raise NessusConnectionError(error)
-        self._set_token(response[TOKEN])
-        self.session = session
+            try:
+                response = session.post(self._get_url_request(SESSION), headers=self.headers,
+                                        json={USERNAME: self.username, PASSWORD: self.password},
+                                        verify=self.verify_ssl, timeout=(5, 30))
+                response.raise_for_status()
+            except requests.HTTPError as e:
+                raise NessusConnectionError(
+                    "POST /session with username {0} failed. Details: {1}".format(self.username, str(e)))
+            response = response.json()
+            if TOKEN not in response:
+                error = response.get('errorCode', 'Unknown connection error')
+                message = response.get('errorMessage', '')
+                if message:
+                    error = '{0}: {1}'.format(error, message)
+                raise NessusConnectionError(error)
+            self._set_token_headers(response[TOKEN])
+            self.session = session
+        else:
+            # In this case we just need to set the headers of the api key
+            self._set_api_headers()
+            # Checking that the connection works
+            self._get(SCANS_ENDPOINT)
 
     def get_scans(self, **kwargs):
         """
@@ -128,6 +153,7 @@ class NessusConnection(object):
 
         :param scan_id: Id of the scan that resulted in requested host
         :param host_id: Id of the host to fetch
+        :param open_session: An open requests session (increase performance)
         :param kwargs: Override _get request parameters
         :return: Expanded details of given host as found by given scan
         """
@@ -142,19 +168,21 @@ class NessusConnection(object):
         Destroy current session. to end the connection
         :return:
         """
-        response = self.session.delete(self._get_url_request(SESSION), headers=self.headers, verify=False)
-        try:
-            response.raise_for_status()
-        except requests.HTTPError as e:
-            raise NessusRequestError("DELETE /session failed. Details: {0}".format(str(e)))
-        self._close()
+        if self.should_use_token:
+            response = self.session.delete(self._get_url_request(SESSION), headers=self.headers, verify=False)
+            try:
+                response.raise_for_status()
+            except requests.HTTPError as e:
+                raise NessusRequestError("DELETE /session failed. Details: {0}".format(str(e)))
+            self._close()
         return True
 
     def _close(self):
         """ Terminate the connection and initialize session and token """
-        self.session.close()
-        self.session = None
-        self.token = None
+        if self.should_use_token:
+            self.session.close()
+            self.session = None
+            self.token = None
 
     def _get_url_request(self, endpoint):
         """ Builds and returns the full url for the request
@@ -164,7 +192,7 @@ class NessusConnection(object):
         """
         return self.url + endpoint
 
-    def _set_token(self, token):
+    def _set_token_headers(self, token):
         """ Sets the API token, in the appropriate header, for valid requests
 
         :param str token: API Token
@@ -172,7 +200,12 @@ class NessusConnection(object):
         self.token = token
         self.headers['X-Cookie'] = 'token={0}'.format(token)
 
-    def _get(self, name, params=None, headers=None):
+    def _set_api_headers(self):
+        """ Setting headers to include the given api key
+        """
+        self.headers['X-ApiKeys'] = f'accessKey={self.access_key}; secretKey={self.secret_key}'
+
+    def _get(self, name, params=None, headers=None, **kwargs):
         """
         Serves a GET request to Nessus API
 
@@ -182,12 +215,22 @@ class NessusConnection(object):
         :return: the response json
         :rtype: dict
         """
-        if not self._is_connected:
+        open_session = kwargs.get("open_session")
+        if not self._is_connected and self.should_use_token:
             raise NessusNotConnected()
         headers = headers if headers is not None else self.headers
         params = params if params is not None else {}
-        response = self.session.get(self._get_url_request(name), params=params,
-                                    headers=headers, verify=self.verify_ssl, timeout=(5, 30))
+        if self.should_use_token:
+            response = self.session.get(self._get_url_request(name), params=params,
+                                        headers=headers, verify=self.verify_ssl, timeout=(5, 30))
+        else:
+            # Using api key instead of session with token
+            if open_session is not None:
+                response = open_session.get(self._get_url_request(name), params=params,
+                                            headers=headers, verify=self.verify_ssl, timeout=(5, 30))
+            else:
+                response = requests.get(self._get_url_request(name), params=params,
+                                        headers=headers, verify=self.verify_ssl, timeout=(5, 30))
         try:
             response.raise_for_status()
         except requests.HTTPError as e:
